Potential merge to ruin all project (implemented read/write/lseek, also have done some refactoring of old code)

+++ FileSystem/file_system/entities/Descriptor.cpp
@@ -12,7 +12,7 @@ Descriptor::~Descriptor() {
 
 void Descriptor::parse(const char *bytes) {
     file_size = Utils::bytesToInt32(bytes);
-    for (size_t i = 0; i < 3; i++) {
+    for (int i = 0; i < 3; i++) {
         blocks_indices[i] = Utils::bytesToInt32(bytes + 4 * (i + 1));
     }
 }
@@ -25,7 +25,7 @@ int Descriptor::getFileSize() const {
     return file_size;
 }
 
-int Descriptor::getBlockIndex(size_t i) const {
+int Descriptor::getBlockIndex(int i) const {
     return blocks_indices[i];
 }
 
@@ -33,20 +33,20 @@ void Descriptor::setFileSize(int new_file_size) {
     file_size = new_file_size;
 }
 
-void Descriptor::setBlockIndex(size_t i, int value) {
+void Descriptor::setBlockIndex(int i, int value) {
     blocks_indices[i] = value;
 }
 
 void Descriptor::copyBytes(char *buffer) const {
     char *file_size_bytes = Utils::intToBytes(file_size);
-    for (size_t i = 0; i < 4; i++) {
+    for (int i = 0; i < 4; i++) {
         buffer[i] = file_size_bytes[i];
     }
     delete file_size_bytes;
 
-    for (size_t i = 0; i < 3; i++) {
+    for (int i = 0; i < 3; i++) {
         char *index_bytes = Utils::intToBytes(blocks_indices[i]);
-        for (size_t j = 0; j < 4; j++) {
+        for (int j = 0; j < 4; j++) {
             buffer[4 * (i + 1) + j] = index_bytes[j];
         }
         delete index_bytes;
